+++ resources/js/components/Card.vue
@@ -1,9 +1,22 @@
 <template>
-    <div class="overflow-x-auto bg-white dark:bg-gray-800 rounded-lg shadow h-full pt-4">
-        <h1 v-if="title" class="h-6 flex mb-3 text-sm font-bold px-6">{{ title }}</h1>
+    <div class="overflow-x-auto bg-white dark:bg-gray-800 rounded-lg shadow h-full">
+        <div v-if="title || useViewAll" 
+            class="flex items-center justify-between py-3 text-sm px-6">
+            <h1 v-if="title" class="font-bold">
+                {{ title }}
+            </h1>
+            <div v-if="useViewAll && !viewAllBottomPosition">
+                <a :class="viewAllUseButton ?
+                        'flex-shrink-0 shadow rounded focus:outline-none ring-primary-200 dark:ring-gray-600 focus:ring bg-primary-500 hover:bg-primary-400 active:bg-primary-600 text-white dark:text-gray-800 inline-flex items-center font-bold px-3 h-8 text-xs' :
+                        'link-default'" 
+                    :href="viewAll.link"
+                >
+                    {{ viewAll.label ?? 'View All' }}
+                </a>
+            </div>
+        </div>
         <table
             class="w-full"
-            :class="card.style"
             data-testid="resource-table"
             ref="table"
         >
@@ -22,10 +35,16 @@
                 />
             </tbody>
         </table>
-        <div v-if="viewAll && viewAll.label" class="w-full border-t border-gray-200 dark:border-gray-700 rounded-b-lg flex justify-center py-3">
-            <div>
-                <a class="text-primary-200 text-xs hover:text-primary-600" :href="viewAll.link">{{ viewAll.label }}</a>
-            </div>
+        <div v-if="useViewAll &&viewAllBottomPosition" 
+            class="w-full border-t border-gray-200 dark:border-gray-700 rounded-b-lg flex justify-center py-3"
+        >
+            <a :class="viewAllUseButton ?
+                        'flex-shrink-0 shadow rounded focus:outline-none ring-primary-200 dark:ring-gray-600 focus:ring bg-primary-500 hover:bg-primary-400 active:bg-primary-600 text-white dark:text-gray-800 inline-flex items-center font-bold px-3 h-8 text-xs' :
+                        'link-default'" 
+                :href="viewAll.link"
+                >
+                {{ viewAll.label ?? 'View All' }}
+            </a>
         </div>
 
         <Pagination v-if="paginator" v-model="paginator" @update-rows="update"></Pagination>
@@ -51,7 +70,7 @@ export default {
             rows: [],
             header: [],
             title: '',
-            viewAll: false,
+            viewAll: null,
             paginator: null,
         }
     },
@@ -61,7 +80,7 @@ export default {
         },
 
         shouldShowTight() {
-            return this.card.style === 'table-tight';
+            return this.card.style === 'tight';
         },
 
         /**
@@ -70,6 +89,15 @@ export default {
         shouldShowColumnBorders() {
             return !! this.card.showBorders;
         },
+        useViewAll() {
+            return this.viewAll && this.viewAll.link;
+        },
+        viewAllUseButton() {
+            return this.viewAll && this.viewAll.style && this.viewAll.style === 'button';
+        },
+        viewAllBottomPosition() {
+            return this.viewAll && this.viewAll.position && this.viewAll.position === 'bottom';
+        }
     },
     methods: {
         update(event){
@@ -77,12 +105,13 @@ export default {
         }
     },
     created () {
-		const {header, rows, title, paginator} = this.card;
+		const {header, rows, title, paginator, viewAll} = this.card;
 
 		this.header = header;
 		this.title = title;
         this.rows = rows;
         this.paginator = paginator;
+        this.viewAll = viewAll;
 	},
 }
 </script>